feat: перенос media на fragment
