+++ app/src/main/AndroidManifest.xml
@@ -21,15 +21,6 @@
             android:exported="false"
             android:screenOrientation="portrait" />
         <activity
-            android:name=".media.ui.MediaActivity"
-            android:exported="false"
-            android:screenOrientation="portrait" />
-        <activity
-            android:name=".search.ui.SearchActivity"
-            android:exported="false"
-            android:screenOrientation="portrait"
-            android:windowSoftInputMode="adjustResize" />
-        <activity
             android:name=".settings.ui.SettingsActivity"
             android:exported="false"
             android:screenOrientation="portrait" />
