--- app/src/main/java/com/example/playlistmaker/main/ui/MainActivity.kt
@@ -1,36 +1,0 @@
-package com.example.playlistmaker.main.ui
-
-import android.content.Intent
-import android.os.Bundle
-import android.widget.Button
-import androidx.appcompat.app.AppCompatActivity
-import com.example.playlistmaker.R
-import com.example.playlistmaker.media.ui.MediaActivity
-import com.example.playlistmaker.search.ui.SearchActivity
-import com.example.playlistmaker.settings.ui.SettingsActivity
-
-class MainActivity : AppCompatActivity() {
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_main)
-
-        val buttonSearch = findViewById<Button>(R.id.search)
-        val buttonMedia = findViewById<Button>(R.id.media)
-        val buttonSettings = findViewById<Button>(R.id.settings)
-
-        buttonSettings.setOnClickListener {
-            val displayIntent = Intent(this, SettingsActivity::class.java)
-            startActivity(displayIntent)
-        }
-
-        buttonSearch.setOnClickListener {
-            val displayIntent = Intent(this, SearchActivity::class.java)
-            startActivity(displayIntent)
-        }
-
-        buttonMedia.setOnClickListener {
-            val displayIntent = Intent(this, MediaActivity::class.java)
-            startActivity(displayIntent)
-        }
-    }
-}
