--- app/src/main/java/com/example/playlistmaker/media/ui/MediaActivity.kt
@@ -1,35 +1,0 @@
-package com.example.playlistmaker.media.ui
-
-import android.os.Bundle
-import androidx.appcompat.app.AppCompatActivity
-import com.example.playlistmaker.R
-import com.example.playlistmaker.databinding.ActivityMediaBinding
-import com.google.android.material.tabs.TabLayoutMediator
-
-class MediaActivity : AppCompatActivity() {
-
-    private lateinit var binding: ActivityMediaBinding
-    private lateinit var tabMediator: TabLayoutMediator
-
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-
-        binding = ActivityMediaBinding.inflate(layoutInflater)
-        setContentView(binding.root)
-
-        binding.meToolbar.setNavigationOnClickListener {
-            onBackPressedDispatcher.onBackPressed()
-        }
-
-        binding.viewPager.adapter = MediaViewPagerAdapter(supportFragmentManager, lifecycle)
-        tabMediator = TabLayoutMediator(binding.tabLayout, binding.viewPager) { tab, position ->
-            if (position == 0) tab.setText(R.string.favorites) else tab.setText(R.string.playlists)
-        }
-        tabMediator.attach()
-    }
-
-    override fun onDestroy() {
-        super.onDestroy()
-        tabMediator.detach()
-    }
-}
