+++ app/src/main/java/com/example/playlistmaker/media/ui/MediaFragment.kt
@@ -1,0 +1,36 @@
+package com.example.playlistmaker.media.ui
+
+import android.os.Bundle
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import androidx.fragment.app.Fragment
+import com.example.playlistmaker.R
+import com.example.playlistmaker.databinding.FragmentMediaBinding
+import com.google.android.material.tabs.TabLayoutMediator
+
+class MediaFragment : Fragment() {
+
+    private lateinit var binding: FragmentMediaBinding
+    private lateinit var tabMediator: TabLayoutMediator
+
+    override fun onCreateView(inflater: LayoutInflater, container: ViewGroup?, savedInstanceState: Bundle?): View {
+        binding = FragmentMediaBinding.inflate(inflater, container, false)
+        return binding.root
+    }
+
+    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
+        super.onViewCreated(view, savedInstanceState)
+
+        binding.viewPager.adapter = MediaViewPagerAdapter(childFragmentManager, lifecycle)
+        tabMediator = TabLayoutMediator(binding.tabLayout, binding.viewPager) { tab, position ->
+            if (position == 0) tab.setText(R.string.favorites) else tab.setText(R.string.playlists)
+        }
+        tabMediator.attach()
+    }
+
+    override fun onDestroy() {
+        super.onDestroy()
+        tabMediator.detach()
+    }
+}
